pytrace: split pytorch into its own event and dump format
PyTorch (RecordFunction) and pytrace (PyEval) shared wpytrace_event and
wpytrace_data_hdr, with the merge selecting behavior on an is_pytorch
flag and pytorch leaving code_key and the code map unused.

Move pytorch to its own pytorch_data.h: wpytorch_event without code_key
and wpytorch_data_hdr without the code-map fields. wpytrace_data_hdr is
now pytrace-only. The merge uses separate wpytrace_state and
wpytorch_state, and persist has persist_pytorch_event; no shared type or
is_pytorch flag remains.

PyTorch records are now 16 bytes instead of 24.

+++ src/inj_pytrace.c
@@ -259,7 +259,7 @@ static int pytrace_profile_callback(PyObject *obj, PyFrameObject *frame, int wha
 	struct wpytrace_event ev = {
 		.ts = ts,
 		.tid = tid,
-		.what = (u8)what,
+		.what = what,
 		.code_key = code_key,
 	};
